+++ src/main/java/com/gdsc/homework/HW5/repository/article/dto/request/ArticleServiceRequest.java
@@ -1,6 +1,5 @@
-package com.gdsc.homework.HW5.service.article.dto.request;
+package com.gdsc.homework.HW5.repository.article.dto.request;
 
-import com.gdsc.homework.HW5.domain.Article;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,9 +9,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class ArticleServiceRequest {
-    Long userId;
-    String title;
-    String content;
+    private Long userId;
+    private String title;
+    private String content;
 
     public static ArticleServiceRequest newInstance(Long userId, String title, String content) {
         return new ArticleServiceRequest(userId, title, content);
